pyproject.toml
[project]
name = "meeting-scribe"
version = "0.1.0"
description = "Record meetings, transcribe and summarize with Gemini AI"
readme = "README.md"
authors = [
    { name = "3ntropy", email = "f.society.disobey@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.70",
    "langgraph>=1.1.4",
    "numpy>=2.4.4",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]

[project.scripts]
meeting-scribe = "meeting_scribe.cli:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"
